change repository-service

+++ src/main/java/com/example/sportstore06/controller/CategoryController.java
@@ -4,10 +4,11 @@ import com.example.sportstore06.dao.request.CategoryRequest;
 import com.example.sportstore06.dao.response.CategoryResponse;
 import com.example.sportstore06.model.Category;
 import com.example.sportstore06.service.CategoryService;
+import com.example.sportstore06.service.ImageService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.query.sqm.InterpretationException;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -22,19 +23,26 @@ import java.util.Optional;
 @RequestMapping("api/v1/category")
 @RequiredArgsConstructor
 public class CategoryController {
-    @Value("${page-number}")
-    private Integer page_number;
+    @Value("${page_size_default}")
+    private Integer page_size_default;
     private final CategoryService categoryService;
+    private final ImageService imageService;
 
-    @GetMapping("/findById")
-    public ResponseEntity<?> findById(@RequestParam(value = "id", required = true) Integer id) {
-        return categoryService.findById(id);
+    @GetMapping("/{id}")
+    public ResponseEntity<?> findById(@PathVariable("id") Integer id) {
+        try {
+            if (categoryService.findById(id).isPresent()) {
+                CategoryResponse p = new CategoryResponse(categoryService.findById(id).get());
+                return ResponseEntity.status(HttpStatus.OK).body(p);
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id category not found");
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
-    @GetMapping("/findAll")
-    public ResponseEntity<?> findAll() {
-        return ResponseEntity.ok(categoryService.findAll().stream().map(category -> new CategoryResponse(category)));
-    }
-    @GetMapping("/findByPage")
+
+    @GetMapping()
     public ResponseEntity<?> findByPage(@RequestParam(value = "page", required = false) Optional<Integer> page,
                                         @RequestParam(value = "page_size", required = false) Optional<Integer> page_size,
                                         @RequestParam(value = "sort", required = false) String sort,
@@ -42,26 +50,64 @@ public class CategoryController {
         try {
             Pageable pageable;
             if (sort != null) {
-                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_number),
+                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_size_default),
                         desc.orElse(true) ? Sort.by(sort).descending() : Sort.by(sort).ascending());
             } else {
-                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_number));
+                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_size_default));
             }
-            return categoryService.findByPage(pageable);
-        } catch (InterpretationException exception) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("filed not found");
+            Page<Category> byPage = categoryService.findByPage(pageable);
+            Page<CategoryResponse> responses = byPage.map(category -> new CategoryResponse(category));
+            return ResponseEntity.status(HttpStatus.OK).body(responses);
+        } catch (InvalidDataAccessApiUsageException exception) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("filed name does not exit");
+        }
+    }
+
+    @PostMapping("/save")
+    private ResponseEntity<?> addCategory(@Valid @RequestBody CategoryRequest request) {
+        try {
+            if (imageService.findById(request.getId_image()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id image not found ");
+            }
+            categoryService.save(0, request);
+            return ResponseEntity.accepted().build();
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         }
     }
 
-    @PostMapping("/save")
-    private ResponseEntity<?> save(@Valid @RequestBody CategoryRequest categoryRequest) {
-        return categoryService.save(categoryRequest);
+    @PutMapping("/save/{id}")
+    private ResponseEntity<?> changeCategory(@Valid @RequestBody CategoryRequest request,
+                                             @PathVariable("id") Integer id) {
+        try {
+            if (categoryService.findById(id).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id category not found ");
+            }
+            if (imageService.findById(request.getId_image()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id image not found ");
+            }
+            categoryService.save(id, request);
+            return ResponseEntity.accepted().build();
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
 
-    @DeleteMapping("/deleteById/{id}")
+    @DeleteMapping("/delete/{id}")
     private ResponseEntity<?> deleteById(@PathVariable("id") Integer id) {
-        return categoryService.deleteById(id);
+        try {
+            if (categoryService.findById(id).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id category not found");
+            } else {
+                boolean checkDelete = categoryService.deleteById(id);
+                if (checkDelete) {
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("can't delete");
+                }
+                return ResponseEntity.accepted().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
+
 }
