+++ src/main/java/com/example/sportstore06/controller/CommentController.java
@@ -1,16 +1,15 @@
 package com.example.sportstore06.controller;
 
 import com.example.sportstore06.dao.request.CommentRequest;
-import com.example.sportstore06.dao.request.ProductRequest;
 import com.example.sportstore06.dao.response.CommentResponse;
-import com.example.sportstore06.dao.response.ProductResponse;
 import com.example.sportstore06.model.Comment;
-import com.example.sportstore06.model.Product;
 import com.example.sportstore06.service.CommentService;
+import com.example.sportstore06.service.ProductService;
+import com.example.sportstore06.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.query.sqm.InterpretationException;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -25,15 +24,27 @@ import java.util.Optional;
 @RequestMapping("api/v1/comment")
 @RequiredArgsConstructor
 public class CommentController {
-    @Value("${page-number}")
-    private Integer page_number;
+    @Value("${page_size_default}")
+    private Integer page_size_default;
     private final CommentService commentService;
+    private final ProductService productService;
+    private final UserService userService;
 
-    @GetMapping("/findById")
-    public ResponseEntity<?> findById(@RequestParam(value = "id", required = true) Integer id) {
-        return commentService.findById(id);
+    @GetMapping("/{id}")
+    public ResponseEntity<?> findById(@PathVariable("id") Integer id) {
+        try {
+            if (commentService.findById(id).isPresent()) {
+                CommentResponse p = new CommentResponse(commentService.findById(id).get(), commentService.findByReply(id).orElse(null));
+                return ResponseEntity.status(HttpStatus.OK).body(p);
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id comment not found");
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
-    @GetMapping("/findByPage")
+
+    @GetMapping()
     public ResponseEntity<?> findByPage(@RequestParam(value = "page", required = false) Optional<Integer> page,
                                         @RequestParam(value = "page_size", required = false) Optional<Integer> page_size,
                                         @RequestParam(value = "sort", required = false) String sort,
@@ -41,26 +52,75 @@ public class CommentController {
         try {
             Pageable pageable;
             if (sort != null) {
-                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_number),
+                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_size_default),
                         desc.orElse(true) ? Sort.by(sort).descending() : Sort.by(sort).ascending());
             } else {
-                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_number));
+                pageable = PageRequest.of(page.orElse(0), page_size.orElse(page_size_default));
             }
-            return commentService.findByPage(pageable);
-        } catch (InterpretationException exception) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("filed not found");
+            Page<Comment> byPage = commentService.findByPage(pageable);
+            Page<CommentResponse> responses = byPage.map(comment -> new CommentResponse(comment, commentService.findByReply(comment.getId()).orElse(null)));
+            return ResponseEntity.status(HttpStatus.OK).body(responses);
+        } catch (InvalidDataAccessApiUsageException exception) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("filed name does not exit");
+        }
+    }
+
+    @PostMapping("/save")
+    private ResponseEntity<?> addComment(@Valid @RequestBody CommentRequest request) {
+        try {
+            if (productService.findById(request.getId_product()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id product not found ");
+            } else if (userService.findById(request.getId_user()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id user not found");
+            } else if (commentService.findById(request.getReply()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id reply not found");
+            } else {
+                commentService.save(0, request);
+                return ResponseEntity.accepted().build();
+            }
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         }
     }
 
-    @PostMapping("/save")
-    private ResponseEntity<?> save(@Valid @RequestBody CommentRequest commentRequest) {
-        return commentService.save(commentRequest);
+    @PutMapping("/save/{id}")
+    private ResponseEntity<?> changeComment(@Valid @RequestBody CommentRequest request,
+                                            @PathVariable("id") Integer id) {
+        try {
+            if (commentService.findById(id).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id comment not found");
+            }
+            if (productService.findById(request.getId_product()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id product not found ");
+            }
+            if (userService.findById(request.getId_user()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id user not found");
+            }
+            if (commentService.findById(request.getReply()).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id reply not found");
+            }
+            commentService.save(id, request);
+            return ResponseEntity.accepted().build();
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
 
-    @DeleteMapping("/deleteById/{id}")
+    @DeleteMapping("/delete/{id}")
     private ResponseEntity<?> deleteById(@PathVariable("id") Integer id) {
-        return commentService.deleteById(id);
+        try {
+            if (commentService.findById(id).isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id comment not found");
+            } else {
+                boolean checkDelete = commentService.deleteById(id);
+                if (checkDelete) {
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("can't delete");
+                }
+                return ResponseEntity.accepted().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
+
 }
